Improve minikube stop
