+++ test/integration/util/minikube_runner.go
@@ -20,7 +20,6 @@ import (
 	"bufio"
 	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"math/rand"
 	"os/exec"
@@ -36,8 +35,6 @@ import (
 	"k8s.io/minikube/pkg/minikube/assets"
 	commonutil "k8s.io/minikube/pkg/util"
 )
-
-const kubectlBinary = "kubectl"
 
 // MinikubeRunner runs a command
 type MinikubeRunner struct {
@@ -256,62 +253,6 @@ func (m *MinikubeRunner) CheckStatusNoFail(desired string) error {
 		return fmt.Errorf("got state: %q, expected %q", s, desired)
 	}
 	return nil
-}
-
-// KubectlRunner runs a command using kubectl
-type KubectlRunner struct {
-	T          *testing.T
-	BinaryPath string
-}
-
-// NewKubectlRunner creates a new KubectlRunner
-func NewKubectlRunner(t *testing.T) *KubectlRunner {
-	p, err := exec.LookPath(kubectlBinary)
-	if err != nil {
-		t.Fatalf("Couldn't find kubectl on path.")
-	}
-	return &KubectlRunner{BinaryPath: p, T: t}
-}
-
-// RunCommandParseOutput runs a command and parses the JSON output
-func (k *KubectlRunner) RunCommandParseOutput(args []string, outputObj interface{}) error {
-	args = append(args, "-o=json")
-	output, err := k.RunCommand(args)
-	if err != nil {
-		return err
-	}
-	d := json.NewDecoder(bytes.NewReader(output))
-	if err := d.Decode(outputObj); err != nil {
-		return err
-	}
-	return nil
-}
-
-// RunCommand runs a command, returning stdout
-func (k *KubectlRunner) RunCommand(args []string) (stdout []byte, err error) {
-	inner := func() error {
-		cmd := exec.Command(k.BinaryPath, args...)
-		stdout, err = cmd.CombinedOutput()
-		if err != nil {
-			retriable := &commonutil.RetriableError{Err: fmt.Errorf("error running command %s: %v. Stdout: \n %s", args, err, stdout)}
-			k.T.Log(retriable)
-			return retriable
-		}
-		return nil
-	}
-
-	err = commonutil.RetryAfter(3, inner, 2*time.Second)
-	return stdout, err
-}
-
-// CreateRandomNamespace creates a random namespace
-func (k *KubectlRunner) CreateRandomNamespace() string {
-	const strLen = 20
-	name := genRandString(strLen)
-	if _, err := k.RunCommand([]string{"create", "namespace", name}); err != nil {
-		k.T.Fatalf("Error creating namespace: %v", err)
-	}
-	return name
 }
 
 func genRandString(strLen int) string {
